init: helm browse notes
I frequently need to consult my notes inside ~/org and
~/org/projects. This commit introduces two new functions
aru/helm-browse-notes and aru/helm-browse-project-notes to do
this. The functions use helm-find-files-1 internally, and are bound to
"C-c n n" and "C-c n p" respectively. One catch is that we need to run
any of the helm-* commands before these commands since helm needs to
be configured by use-package first.