init: org-ref
I tried out org-ref as an alternative to Zotero to manage citations
and notes for the literature that I read. The package is too powerful
for my very simple use case right now (keep all papers in a single
location, and collect notes on them). The package is disabled for now.