org-zettel: init org zettel
This file is for the slipbox method of Luhmann, replicated in org. One
of the things that we need to do when using the slipbox is to see all
notes that link to a note (say note A). When using a physical slipbox
with index cards, this is easy to do, simply group the cards together
and arrange them in a surface.

To mimic this behaviour, this commit introduces the
~org-zettel-get-inbound-links~ function which shows all notes that
link to the current note. The title of the current note that the
cursor is on is passed on to ~org-occur~.