init: disable tramp settings
Faced a lot of issues with TRAMP for multihop scenarios (which is my
primary use case). Since I have resorted to using shell scripts to
rsync files between my machines, I can disable this for now since it
takes a long time to load (0.22 seconds). This is suspecious however I
do not have the time to investigate this at the moment.