[feature] custom wrapper around load-theme
Emacs deals with themes a bit differently, in the sense that it allows
you to enable several themes at once (after all, themes are simply
adjustments to various "faces" which can be overriden). Needless to
say, constantly using =load-theme= to enable new themes is a bad idea
since all the themes start to overlap and things look weird. The
solution is to first disable the current theme using =disable-theme=
and them load the new theme using =load-theme=. This commit introduces
a new function =aru/load-theme= which does exactly this. The user is
first prompted to interactively pick a new theme following which the
current theme is disabled and the new theme loaded.