[feature] init recentf and smooth scrolling
recentf is a package which enables emacs to keep track of recent
files. This commit enables the recentf minor mode along with declaring
=aru/recentf-find-file= which works just like =find-file= but for
recently opened files only. The function is bound to =C-x C-r= which
by default is bound to ffap-read-only (note I have ffap-mode turned
on) however I have used to use this feature so I am okay with the
rebind for now.