[feature] use the right python
Homebrew provides the =usr/local/opt/python@x.x/libexec/bin= directory
which contains unversioned executables for python and pip which are
correctly symlinked to the actual python@x.x executables.

This commit adds this directory to Emacs's path so that the correct
version of python is picked up when we run =M-x run-python= (in this
case it happens to be python@3.8).

Although this approach requires me to update the path with every new
release of python, imo it is the simplest fix and avoids the
requirement of external tools such as pyenv (which manipulates the
path anyway) and exec-shell-from-path (which will never be as fast as
using =setenv=).