[refactor] setup org-publish for website
I am using org-publish to generate my website the source for which is
simply a bunch of org files in =~/org= directory. I exclude a few
files namely: =inbox.org=, =org-agenda-files.org= and all files that
begin with the word =private=. Additionally, I add a meta tag to the
html files which prevent web crawlers from indexing the pages.