[feature] init consult.el package
Introduce minimal config for the consult.el package. I am starting
slow and using it only for search, imenu and buffer navigation. This
commit also removes the older config which is made redundant by the
introduction of this package (such as my custom =aru/recentf-*=
functions as =consult-buffer= includes this functionality).