[fix] org-publish recursive bug
The docs/ folder was getting published twice such that we had
something like docs/docs/assets showing up in the directory tree. The
issue was due to the `:recursive t` flag in set in the org-static and
website-static projects. The fix is to be more specific about the
`:base-directory` and `:publishing-directory`. The fix was tested and
works as expected.

Additionally, for the website, since the CNAME does not exist within
the new base-directory, a separate publishing project had to be
created which only targets and publishes the CNAME file within docs/.
This was also tested and works as expected.