[feature] improved org-capture template for papers
My previous capture template was a bit too manual. Assuming that the
bibkey was in my kill-ring, the capture template would create a TODO
with the header set to the bibkey along with a link to the pdf on the
local filesystem. After this, I had to copy the entire bibtex entry
and save it manually. Finally, I had to save the pdf with the matching
bibkey name such that the link created earlier would work.

In this commit, I make this process a bit more automated. With the
entire bibtex entry copied to the clipboard (which means it is also in
the kill-ring), the template now does the following:

1. extract the bibkey and sets it as the title
2. link to the file on local filesystem
3. put the bibtex entry within a src block