[refactor] aru/org-capture-template-paper
Move the function into a separate org package `aru-org'. Modify
aru/org-capture-template-paper such that now we can also set the
bibkey as the latest item in the kill-ring (because immediately after
capturing in emacs, I save the pdf version of the paper with the
bibkey as the file name). Also modify the template a bit to add the
usual fields I populate while taking notes.