[feature] improve org-capture for papers
In 58a803a I initialised a bit of automation for capturing
bibliographic info into org mode. With this commit, I expand the idea
into an independent emacs package with additional functionalities. The
package now does the following:

+ capture the bibkey and push it into the kill-ring
+ capture the first author
+ capture the last author
+ capture the publication source

The motivation for this extension is to utilize the full power of
org properties and quickly shortlist relevant papers that I am looking
for.

There is a noticable (~1sec) delay when this capture template runs, in
the future I might consider looking into the runtime performance.