[refactor,feature] aru-window: tmux zoom-pane like functionality
I like the zoom-pane feature of tmux. I usually have source code and a
terminal window. After running some command in the terminal window, I
like to zoom-in to see the output. Once time, I like to return to my
previous window. The zoom-in funtionality is already provided by
aru-window minor-mode (this commit renames it to aru-window-mode to
keep it consistent with elisp standards). Additionally, this commit
adds the ability to zoom-out and return to the previous window. This
is done by the `aru-window-other-window-dwim' function.

Additionally, this commit also introduces
`aru-window-split-window-right' and `aru-window-split-window-below'.
These functions act as wrapper around the respective builtin functions
with some added functionality such as switching to the new window and
balancing the windows.