[refactor] outline-minor-mode
outline-minor-mode in Emacs@28 has become smarter! When enabled in
prog-mode buffers, it handles visiblity toggling of the comment
headers and code blocks! Previously I was using a combination of
outline-minor-mode and hs-minor-mode along with the bicycle package to
glue their functionality together, across {text,prog}-mode buffers.
This is no longer needed so in this commit I remove my old config.