[refactor] simplify aru-window
Previously, I was using a minor-mode to indicate that I was zoomed
into a window with a custom keybinding (=s-m= to toggle zoom). This
commit simplifies aru-window by removing the minor mode. Instead,
=s-o= is bound to `aru-window-other-window-dwim' which performs a
regular `other-window' when more than 1 window is open in a frame. If
only 1 window is visible, then it performs a `winner-undo' before
`other-window'.