[feature] init aru-narrow.el
I use the narrow functionality in emacs daily, specially when I want
to focus on a particular org subtree or edit a particular paragraph. I
am not a big fan of the default keybindings though which require me to
issue separate keys to narrow to a particular
function/paragraph/subtree/region and then another key to widen the
buffer.

This commit introduces the /aru-narrow.el/ package which introduces a
single function =aru-narrow-or-widen-dwim=. This is adapted from the
gist available at
<https://gist.github.com/mwfogleman/95cc60c87a9323876c6c>. The
function does the following:

+ if the current buffer is already narrowed, widen it
+ else if a region is active, narrow to region
+ else if in a org-mode buffer, narrow to subtree
+ as a last resort, call =narrow-to-defun= (which works most of the
time in any major-mode)

I bind this function to =C-x n n= (default is narrow-to-region).